+++ DataModels/WorkoutInterval.cs
@@ -17,11 +17,11 @@
 
         public int? IntervalNo { get; set; }
 
-        public float TimeOffset { get; set; }
+        public double TimeOffset { get; set; }
 
-        public float Duration { get; set; }
+        public double Duration { get; set; }
 
-        public float Distance { get; set; }
+        public double Distance { get; set; }
 
         //[NotMapped]
         //public float DurationComputed { get => Lengths?.Sum(x => x.Duration) ?? 0; }
@@ -36,15 +36,15 @@
         [MaxLength(1000)]
         public string Notes { get; set; }
 
-        public float StrokeCount { get; set; }
+        public double StrokeCount { get; set; }
 
         public WorkoutIntervalType WorkoutIntervalType { get; set; }
 
         //[DatabaseGenerated(DatabaseGeneratedOption.Computed)] // this is set in OnModelCreating() so it is redundant here
         public StrokeType? StrokeTypeId { get; private set; }
-        public float? Pace { get; set; }
+        public double? Pace { get; set; }
 
-        public float? Swolf { get; set; }
+        public double? Swolf { get; set; }
 
         public virtual Workout Workout { get; set; }
     }
